Factor out API v2 into it’s own directory.
