+++ openspending/views/api_v2/__init__.py
@@ -1,0 +1,7 @@
+# from openspending.views.api_v2.aggregate import aggregate # noqa
+# from openspending.views.api_v2.search import search # noqa
+# from openspending.views.api_v2.new import create # noqa
+
+# from openspending.views.api_v2.common import blueprint # noqa
+
+from openspending.views.api_v2.aggregate import blueprint
