+++ openspending/views/api_v2/common.py
@@ -1,0 +1,3 @@
+from flask import Blueprint
+
+blueprint = Blueprint('api_v2', __name__)
